Separate creation DB part for now. (Until we refactor properly)
